requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-cache2==0.2.1
fastapi-users-db-sqlalchemy==6.0.1
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
orjson==3.9.15
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pwdlib==0.2.0
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
time-machine==2.14.0
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
